requirements.txt
channels==2.1.7
channels-redis==2.3.3
Django==2.1.4
django-cors-headers==2.4.0
django-mysql==2.4.1
djangorestframework==3.9.0
djangorestframework-simplejwt==4.0.0
drf-tracking==1.5.0
nose==1.3.7
mysql-connector-python==8.0.15
pytest-asyncio==0.10.0
pytest-django==3.4.7
requests==2.21.0
sentry-sdk==0.6.2
uWSGI==2.0.17.1
whitenoise==4.1.2
